pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pa-reply"
version = "0.0.1"
description = "Passive-Aggressive-Reply-as-a-Service is a simple API that returns passive-aggressive replies."
authors = [{ name = "Diogo Ferreira" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"

dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "huggingface_hub",
  "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["pa_reply"]
force_sort_within_sections = true
